+++ packages/lumx-react/src/components/radio-button/RadioButton.stories.tsx
@@ -1,6 +1,8 @@
 import { RadioButton } from '@lumx/react';
 import { withValueOnChange } from '@lumx/react/stories/decorators/withValueOnChange';
 import { loremIpsum } from '@lumx/react/stories/utils/lorem';
+import { withCombinations } from '@lumx/react/stories/decorators/withCombinations';
+import uniqueId from 'lodash/uniqueId';
 
 export default {
     title: 'LumX components/radio-button/Radio button',
@@ -36,4 +38,34 @@ export const LabelAndHelper = {
         label: 'Radio button label',
         helper: loremIpsum('tiny'),
     },
+};
+
+/**
+ * All state combinations
+ */
+export const AllStates = {
+    args: { ...LabelAndHelper.args, helper: 'Radio button helper' },
+    decorators: [
+        withCombinations({
+            combinations: {
+                rows: {
+                    Default: {},
+                    Checked: { isChecked: true },
+                },
+                cols: {
+                    Default: {},
+                    Disabled: { isDisabled: true },
+                    'ARIA Disabled': { 'aria-disabled': true },
+                },
+            },
+            combinator(a, b) {
+                return Object.assign(a, b, {
+                    // Injecting a unique name for each radio buttons to make sure they can be individually focused
+                    name: uniqueId('name'),
+                    // Disabling
+                    onChange: undefined,
+                });
+            },
+        }),
+    ],
 };