+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Chip`: handle `aria-disabled=true` for an accessible disabled state
 -   `Link`: handle `aria-disabled=true` for an accessible disabled state
 -   `ListItem`: handle `aria-disabled=true` for an accessible disabled state
+-   `RadioButton`: handle `aria-disabled=true` for an accessible disabled state
 
 ### Fixed